+++ lib/helpers/Format.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 namespace yii\mustache\helpers;
 
-use yii\helpers\{Html as HTMLHelper};
+use yii\helpers\{Html as HtmlHelper};
 use yii\mustache\{Helper};
 
 /**
@@ -24,8 +24,8 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as boolean.
    */
   public function getBoolean(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
-      return HTMLHelper::encode(\Yii::$app->formatter->asBoolean($helper->render($value)));
+    return function($value, \Mustache_LambdaHelper $helper) {
+      return HtmlHelper::encode(\Yii::$app->formatter->asBoolean($helper->render($value)));
     };
   }
 
@@ -34,14 +34,14 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as currency number.
    */
   public function getCurrency(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', [
         'currency' => null,
         'options' => [],
         'textOptions' => []
       ]);
 
-      return HTMLHelper::encode(\Yii::$app->formatter->asCurrency(
+      return HtmlHelper::encode(\Yii::$app->formatter->asCurrency(
         $args['value'],
         $args['currency'],
         $args['options'],
@@ -55,9 +55,9 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as date.
    */
   public function getDate(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', ['format' => null]);
-      return HTMLHelper::encode(\Yii::$app->formatter->asDate($args['value'], $args['format']));
+      return HtmlHelper::encode(\Yii::$app->formatter->asDate($args['value'], $args['format']));
     };
   }
 
@@ -66,9 +66,9 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as datetime.
    */
   public function getDateTime(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', ['format' => null]);
-      return HTMLHelper::encode(\Yii::$app->formatter->asDatetime($args['value'], $args['format']));
+      return HtmlHelper::encode(\Yii::$app->formatter->asDatetime($args['value'], $args['format']));
     };
   }
 
@@ -77,14 +77,14 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as decimal number.
    */
   public function getDecimal(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', [
         'decimals' => null,
         'options' => [],
         'textOptions' => []
       ]);
 
-      return HTMLHelper::encode(\Yii::$app->formatter->asDecimal(
+      return HtmlHelper::encode(\Yii::$app->formatter->asDecimal(
         $args['value'],
         $args['decimals'],
         $args['options'],
@@ -98,13 +98,13 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as integer number without rounding.
    */
   public function getInteger(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', [
         'options' => [],
         'textOptions' => []
       ]);
 
-      return HTMLHelper::encode(\Yii::$app->formatter->asInteger(
+      return HtmlHelper::encode(\Yii::$app->formatter->asInteger(
         $args['value'],
         $args['options'],
         $args['textOptions']
@@ -117,9 +117,9 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as HTML-encoded text with newlines converted into breaks.
    */
   public function getNtext(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       if (!isset($value)) return \Yii::$app->formatter->nullDisplay;
-      return preg_replace('/\r?\n/', '<br>', HTMLHelper::encode($helper->render($value)));
+      return preg_replace('/\r?\n/', '<br>', HtmlHelper::encode($helper->render($value)));
     };
   }
 
@@ -128,14 +128,14 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as percent number.
    */
   public function getPercent(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', [
         'decimals' => null,
         'options' => [],
         'textOptions' => []
       ]);
 
-      return HTMLHelper::encode(\Yii::$app->formatter->asPercent(
+      return HtmlHelper::encode(\Yii::$app->formatter->asPercent(
         $args['value'],
         $args['decimals'],
         $args['options'],
@@ -149,9 +149,9 @@ class Format extends Helper {
    * @return \Closure A function formatting a value as time.
    */
   public function getTime(): \Closure {
-    return function($value, \Mustache_LambdaHelper $helper): string {
+    return function($value, \Mustache_LambdaHelper $helper) {
       $args = $this->parseArguments($helper->render($value), 'value', ['format' => null]);
-      return HTMLHelper::encode(\Yii::$app->formatter->asTime($args['value'], $args['format']));
+      return HtmlHelper::encode(\Yii::$app->formatter->asTime($args['value'], $args['format']));
     };
   }
 }