+++ test/LoaderTest.php
@@ -20,8 +20,7 @@ class LoaderTest extends TestCase {
    * @test Loader::findViewFile
    */
   public function testFindViewFile() {
-    $findViewFile = function(string $name): string {
-      /** @var Loader $this */
+    $findViewFile = function($name) {
       return $this->findViewFile($name);
     };